+++ sdk/communication/azure-communication-identity/azure/communication/identity/_generated/aio/_communication_identity_client.py
@@ -9,9 +9,10 @@
 from copy import deepcopy
 from typing import Any, Awaitable
 
+from msrest import Deserializer, Serializer
+
 from azure.core import AsyncPipelineClient
 from azure.core.rest import AsyncHttpResponse, HttpRequest
-from msrest import Deserializer, Serializer
 
 from .. import models
 from ._configuration import CommunicationIdentityClientConfiguration
@@ -26,7 +27,7 @@ class CommunicationIdentityClient:
     :param endpoint: The communication resource, for example
      https://my-resource.communication.azure.com.
     :type endpoint: str
-    :keyword api_version: Api Version. The default value is "2022-06-01". Note that overriding this
+    :keyword api_version: Api Version. Default value is "2022-06-01". Note that overriding this
      default value may result in unsupported behavior.
     :paramtype api_version: str
     """