+++ sdk/communication/azure-communication-identity/azure/communication/identity/_generated/aio/_configuration.py
@@ -13,15 +13,17 @@ from azure.core.pipeline import policies
 
 VERSION = "unknown"
 
-class CommunicationIdentityClientConfiguration(Configuration):
+class CommunicationIdentityClientConfiguration(Configuration):  # pylint: disable=too-many-instance-attributes
     """Configuration for CommunicationIdentityClient.
 
     Note that all parameters used to create this instance are saved as instance
     attributes.
 
-    :param endpoint: The communication resource, for example https://my-resource.communication.azure.com.
+    :param endpoint: The communication resource, for example
+     https://my-resource.communication.azure.com.
     :type endpoint: str
-    :keyword api_version: Api Version. The default value is "2022-06-01". Note that overriding this default value may result in unsupported behavior.
+    :keyword api_version: Api Version. Default value is "2022-06-01". Note that overriding this
+     default value may result in unsupported behavior.
     :paramtype api_version: str
     """