+++ sdk/communication/azure-communication-identity/azure/communication/identity/_generated/models/_communication_identity_client_enums.py
@@ -17,4 +17,3 @@ class CommunicationTokenScope(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)
 
     CHAT = "chat"
     VOIP = "voip"
-    EMAIL = "email"